mobile views, navbar links, about, and faq page fixed
fixed links for navbar, then centered about page, and added faq page. for faq added a sample drop down menu that needs to be desperately better styled. another thing that needs to be fixed is page spacing, really not optimal rn. and also movbile views are updated, so should be seamless as well. left comments on changes and things i noted that need to be updated